Use ImageMagick for Debbugs logo conversion instead of Inkscape
Fix the build that was failing on Inkscape command:

    Unknown option --export-png

This was because parameter was deprecated as documented at
https://wiki.inkscape.org/wiki/Using_the_Command_Line#Deprecations_and_Replacements

However, switching to `--export-type=png` would not have been enough.
Additionally one would also need to wrap the command in `xvfb-run` and
configure both Mesa and ZINK options to satifsy Inkscape requirements
to get past the error `MESA: error: ZINK: failed to choose pdev`.

As a much more elegant solution, simply switch to use ImageMagick to
convert the file from svg to png, as it works well purely from the
command-line and requires no new build dependencies.

Also document how the Makefile for the logo works.

# Conflicts:
#	html/logo/Makefile